backend/requirements.txt
# --- Core Framework ---
fastapi>=0.112.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
pydantic-settings>=2.4.0
python-multipart>=0.0.9
python-dotenv>=1.0.1

# --- Base de Datos ---
pymongo>=4.8.0

# --- IA y LangChain (Versiones Críticas) ---
# Es vital que langchain-google-genai sea la última versión para corregir el error 404
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-google-genai>=2.0.1
langchain-groq>=0.2.0

# Librería base de Google (necesaria para la conexión segura)
google-generativeai>=0.8.3

# --- Herramientas de Validación (Fuzzy Match) ---
thefuzz>=0.22.1
python-Levenshtein>=0.25.1

# --- Utilidades Opcionales ---
requests>=2.32.0
# Pillow es recomendado si LangChain necesita procesar imágenes internamente
Pillow

# --- Embeddings (Opcionales para búsqueda futura) ---
cohere>=5.6.0